reorganized images folder
