+++ src/CoolBytes.WebAPI/Features/Authors/ExperienceViewModel.cs
@@ -1,4 +1,4 @@
-﻿using CoolBytes.WebAPI.Features.Images;
+﻿using CoolBytes.WebAPI.Features.Images.ViewModels;
 
 namespace CoolBytes.WebAPI.Features.Authors
 {
