+++ src/CoolBytes.WebAPI/Features/BlogPosts/Profiles/BlogPostViewModelProfile.cs
@@ -1,7 +1,7 @@
 ﻿using AutoMapper;
 using CoolBytes.Core.Domain;
 using CoolBytes.WebAPI.Features.BlogPosts.ViewModels;
-using CoolBytes.WebAPI.Features.Images;
+using CoolBytes.WebAPI.Features.Images.ViewModels;
 
 namespace CoolBytes.WebAPI.Features.BlogPosts.Profiles
 {
