+++ src/CoolBytes.WebAPI/Features/BlogPosts/ViewModels/BlogPostUpdateViewModel.cs
@@ -1,5 +1,5 @@
 ﻿using CoolBytes.WebAPI.Features.BlogPosts.DTO;
-using CoolBytes.WebAPI.Features.Images;
+using CoolBytes.WebAPI.Features.Images.ViewModels;
 using System;
 using System.Collections.Generic;
