+++ src/CoolBytes.WebAPI/Features/Images/CQ/GetImagesQuery.cs
@@ -1,7 +1,8 @@
 ﻿using System.Collections.Generic;
+using CoolBytes.WebAPI.Features.Images.ViewModels;
 using MediatR;
 
-namespace CoolBytes.WebAPI.Features.Images
+namespace CoolBytes.WebAPI.Features.Images.CQ
 {
     public class GetImagesQuery : IRequest<IEnumerable<ImageViewModel>>
     {
