+++ src/CoolBytes.WebAPI/Features/Images/CQ/UploadImagesCommand.cs
@@ -1,8 +1,9 @@
 ﻿using System.Collections.Generic;
+using CoolBytes.WebAPI.Features.Images.ViewModels;
 using MediatR;
 using Microsoft.AspNetCore.Http;
 
-namespace CoolBytes.WebAPI.Features.Images
+namespace CoolBytes.WebAPI.Features.Images.CQ
 {
     public class UploadImagesCommand : IRequest<IEnumerable<ImageViewModel>>
     {
