+++ src/CoolBytes.WebAPI/Features/Images/Profiles/ImageViewModelProfile.cs
@@ -1,7 +1,8 @@
 ﻿using AutoMapper;
 using CoolBytes.Core.Domain;
+using CoolBytes.WebAPI.Features.Images.ViewModels;
 
-namespace CoolBytes.WebAPI.Features.Images
+namespace CoolBytes.WebAPI.Features.Images.Profiles
 {
     public class ImageViewModelProfile : Profile
     {
