+++ src/CoolBytes.WebAPI/Features/Images/Validators/UploadImageCommandValidator.cs
@@ -1,6 +1,7 @@
-﻿using FluentValidation;
+﻿using CoolBytes.WebAPI.Features.Images.CQ;
+using FluentValidation;
 
-namespace CoolBytes.WebAPI.Features.Images
+namespace CoolBytes.WebAPI.Features.Images.Validators
 {
     public class UploadImageCommandValidator : AbstractValidator<UploadImagesCommand>
     {
